requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
sendgrid==6.12.5
python-dotenv==1.0.1
pydantic>=2.3.0
pydantic-settings>=2.1.0
httpx==0.24.1
pytest==8.4.2
bcrypt==4.1.2
